% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_pam.R
\name{base_PAM}
\alias{base_PAM}
\title{Prepare a presence-absence matrix (PAM)}
\usage{
base_PAM(data, format = NULL, master_matrix, cell_size,
         complete_cover = TRUE)
}
\arguments{
\item{data}{species geographic ranges to be used to create a presence-absence
matrix (PAM). This argument can be: RasterStack, RasterBrick, data.frame,
list, SpatialPolygonsDataFrame, SpatialPointsDataFrame, or character. See
details for description of characteristics of each option.}

\item{format}{(character) if \code{data} is a character, available formats are:
"shp", "gpkg", "GTiff", and "ascii".}

\item{master_matrix}{object of class "master_matrix" or "master_selection". See
details.}

\item{cell_size}{(numeric) resolution for grid (single number or vector of two
numbers) in kilometers (km).}

\item{complete_cover}{(logical) whether or not to include cells of grid
partially overlapped with the geographic region of interest contained in
\code{master_matrix}. Default = TRUE.}
}
\value{
A presence-absence matrix (PAM) of class base_PAM for the region of interest
associated with a SpatialPolygonsDataFrame, as in a grid of \code{cell_size}
resolution. Each grid cell is related to a specific ID and longitude and
latitude coordinates. Presence (1) and absence (0) values for each species
in every cell of the PAM are included as apart of the data frame of the
SpatialPolygonsDataFrame.
}
\description{
Prepare a presence-absence matrix (PAM) in which all initial points
of interest (rows) will have a representation of the species present in such
areas (columns). Initial points of interest will be represented by an ID, and
longitude and latitude coordinates.
}
\details{
Objects of class "master_matrix" or "master_selection" could be obtained from
functions \code{\link{master_matrix}}, \code{\link{random_selection}},
\code{\link{uniformG_selection}}, \code{\link{uniformE_selection}}, or
\code{uniformEG_selection}.
}
\examples{
# Data
data("m_matrix", package = "biosurvey")
data("species_data", package = "biosurvey")

# Create base_PAM
b_pam <- base_PAM(data = species_data, master_matrix = m_matrix, cell_size = 100)
sp::plot(b_pam$PAM)
summary(b_pam$PAM@data[, 1:6])
}
