% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosurvey.R
\docType{package}
\name{biosurvey}
\alias{biosurvey}
\title{biosurvey: Tools for Biological Survey Planning}
\description{
nichevol is a collection of tools that allow users to select sampling sites
increasing efficiency of biodiversity survey systems by considering the
relationship of environmental and geographic conditions in a region.
}
\section{Main functions in biosurvey}{

\code{\link{master_matrix}}, \code{\link{base_pam}}

Other functions (important helpers)

\code{\link{files_2data}}, \code{\link{grid_from_region}},
\code{\link{pam_from_table}}, \code{\link{rlist_2data}},
\code{\link{spdf_2data}}, \code{\link{stack_2data}}
}

