% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_sample.R
\name{block_sample}
\alias{block_sample}
\title{Selection of blocks in environmental space}
\usage{
block_sample(master, variable_1, variable_2, expected_blocks,
             selection_type = "uniform", initial_distance = NULL,
             increase = NULL, replicates = 10, set_seed = 1)
}
\arguments{
\item{master}{a master_matrix object derived from the function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{random_selection}}, \code{\link{uniformG_selection}},
or \code{\link{uniformE_selection}}.}

\item{variable_1}{(character or numeric) name or position of the first
variable (X axis).}

\item{variable_2}{(character or numeric) name or position of the second
variable (Y axis) to be used to create blocks (must be different from the
first one).}

\item{expected_blocks}{(numeric) number of blocks to be selected.}

\item{selection_type}{(character) Type of selection. Two options are available:
"uniform" and "random". Default = "uniform".}

\item{initial_distance}{(numeric) euclidean distance to be used for a first
process of thinning and detection of remaining points. If \code{selection_type}
= "uniform", this argument must be defined. Default = NULL.}

\item{increase}{(numeric) value to be added to \code{initial_distance} until
reaching the number of \code{expected_points}. If \code{selection_type}
= "uniform", this argument must be defined. Default = NULL.}

\item{replicates}{(numeric) number of thinning replicates performed to select
blocks uniformly. Default = 10.}

\item{set_seed}{(numeric) integer value to specify a initial seed. Default = 1.}
}
\value{
An S3 object of class master_matrix or master_selection, containing the same
elements found in the input object, with an additional column in the master_matrix
data.frame containing a binary code for selected (1) and non-selected (0) blocks.
}
\description{
select a user defined number of blocks in environmental space
to be used in further analysis in order to define sampling sites for a survey
system.
}
\details{
When blocks in \code{master} were defined using the option "equal_points"
(see \code{\link{make_blocks}}), "uniform" \code{selection_type} could result
in blocks with high density per area being overlooked.
}
\examples{
# Data
data("m_matrix", package = "biosurvey")

# Making blocks for analysis
m_blocks <- make_blocks(m_matrix, variable_1 = "PC1",
                        variable_2 = "PC2", n_cols = 10, n_rows = 10,
                        block_type = "equal_area")

# Checking column names and values in variables to define initial distance
colnames(m_blocks$master_matrix)
summary(m_blocks$master_matrix[, c("PC1", "PC2")])

# Selecting Blocks uniformly in E space
block_sel <- block_sample(m_blocks, variable_1 = "PC1", variable_2  = "PC2",
                          expected_blocks = 10, selection_type = "uniform",
                          initial_distance = 1.5, increase = 0.1)

head(block_sel$master_matrix)
}
