% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_sites_DI.R
\name{dis_loop}
\alias{dis_loop}
\title{Helper to calculate dissimilarities in loop}
\usage{
dis_loop(site_spp_list, icol, fcol, method = "jaccard", verbose = TRUE, ...)
}
\arguments{
\item{site_spp_list}{list of presence absence matrices for a set of sites or
cells of a grid.}

\item{icol}{number of column where the species list (other columns) starts.}

\item{fcol}{number of column where the species list (other columns) ends.}

\item{method}{(character) dissimilarity index to be passed to function
\code{\link[vegan]{vegdist}}. Default = "jaccard". See details.}

\item{verbose}{(logical) whether or not to print messages about the process.
Default = TRUE.}

\item{...}{other arguments to be passed to function
\code{\link[vegan]{vegdist}}.}
}
\value{
A list of results from \code{\link[vegan]{vegdist}}.
}
\description{
Helper to calculate dissimilarities in loop
}
