% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{dist_list}
\alias{dist_list}
\title{A list of vectors of distances}
\format{A list.
\describe{
\item{vector}{numeric, values of distances (six elements)}
}}
\usage{
dist_list
}
\description{
A list of six vectors of point distances.
}
\examples{
data("dist_list", package = "biosurvey")
}
\keyword{datasets}
