% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionEG_helpers.R
\name{distance_filter}
\alias{distance_filter}
\title{Helper to filter sets of sites by median distance among all points}
\usage{
distance_filter(site_list, median_distance_filter = "max")
}
\arguments{
\item{site_list}{list of selected sites provided as data.frames. Columns
"Longitude" and "Latitude" are needed for distance calculation.}

\item{median_distance_filter}{(character) optional argument to define a median
distance-based filter based on which sets of sampling sites will be selected.
The default, NULL, does not apply such a filter. Options are: "max" and "min".}
}
\description{
Helper to filter sets of sites by median distance among all points
}
