% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data_EG.R
\name{explore_data_EG}
\alias{explore_data_EG}
\title{Plots to explore environmental factors in environmental and geographic space}
\usage{
explore_data_EG(master, variable_1, variable_2, col_variable1 = NULL,
                col_variable2 = NULL, col_points = NULL, col_density = NULL)
}
\arguments{
\item{master}{a master_matrix object derived from function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{uniformG_selection}}, \code{\link{uniformE_selection}}
or \code{EG_selection}.}

\item{variable_1}{(character or numeric) name or position of the first
variable (X axis) to be explored.}

\item{variable_2}{(character or numeric) name or position of the second
variable (Y axis) to be explored (must be different from the first one).}

\item{col_variable1}{a color palette for \code{variable_1} defined
using functions like \code{\link{heat.colors}}, or one generated
using functions like \code{\link{colorRampPalette}}. The default,
NULL, uses \code{viridis::cividis(255)}.}

\item{col_variable2}{a color palette for \code{variable_2} defined
using functions like \code{\link{heat.colors}}, or one generated
using functions like \code{\link{colorRampPalette}}. The default,
NULL, uses \code{viridis::cividis(255)}.}

\item{col_points}{color for points in environmental space. The default, NULL,
uses the 25th color of the default palette for \code{col_variable1} with an
alpha of 0.6.}

\item{col_density}{a color palette to represent representation density of points
in environmental space. This palette can be defined using functions like
\code{\link{heat.colors}}, or one generated using functions like
\code{\link{colorRampPalette}}. The default, NULL, uses
\code{viridis::viridis(255)}, and changes the first color in the palette by NA.}
}
\value{
A multi-panel plot showing two of the environmental predictors in the region of
interest in both spaces, geographic and environmental.
}
\description{
Creates a four-panel plot with information of two environmental
predictors (at a time) in the region of interest. The two top panels contain
the information in geographic space (one predictor per panel). The two panels
at the bottom contain information in a 2D environmental space for the two
variables.
}
\examples{
# Data
data("m_matrix", package = "biosurvey")

colnames(m_matrix$master_matrix)

# Plot
explore_data_EG(m_matrix, variable_1 = "Mean_temperature",
                variable_2 = "Annual_precipitation")
}
