% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionEG_helpers.R
\name{find_modes}
\alias{find_modes}
\title{Find modes in a multimodal distribution}
\usage{
find_modes(density)
}
\arguments{
\item{density}{an object of class density obtained using the function
\code{\link{density}}.}
}
\value{
A data.frame conaining the values corresponding to the modes and the density
for those partiular values.
}
\description{
Find modes in a multimodal distibution of values based on the
desisty of such values.
}
