% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_helpers.R
\name{is_pair}
\alias{is_pair}
\title{Test wheather a number is pair}
\usage{
is_pair(x)
}
\arguments{
\item{x}{(numeric) value to be tested.}
}
\value{
Logical value
}
\description{
Test wheather a number is pair
}
\examples{
is_pair(4)
is_pair(5)
}
