% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{m_matrix_pre}
\alias{m_matrix_pre}
\title{Example of a master_matrix object containing preselected sites}
\format{
A list of 6 elements:
\describe{
\item{data_matrix}{data.frame with 6276 rows and 10 columns}
\item{preselected_sites}{data.frame with 5 rows and 11 columns}
\item{region}{object of class SpatialPolygons*}
\item{mask}{NULL}
\item{raster_base}{object of class SpatialPolygonsDataFrame}
\item{PCA_results}{list of length 5}
}
}
\usage{
m_matrix_pre
}
\description{
A S3 object of class master_matrix.
See function \code{\link{prepare_master_matrix}}.
}
\examples{
data("m_matrix_pre", package = "biosurvey")

print(m_matrix_pre)
}
\keyword{datasets}
