% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{master_matrix}
\alias{master_matrix}
\alias{new_master_matrix}
\title{Constructor of S3 objects of class master_matrix}
\usage{
new_master_matrix(data_matrix, preselected_sites = NULL, region,
                  mask = NULL, raster_base, PCA_results = NULL)
}
\arguments{
\item{data_matrix}{date.frame with information about geographic location of
raster cells, initial environmental data, and if available, the first two
principal components derived from an analysis done with environmental values.}

\item{preselected_sites}{data.frame containing sites that must be included
in posterior selections of sites for the survey system. Columns must be:
"Sites", "Longitude", "Latitude", in that order. Default = NULL.}

\item{region}{SpatVector object representing the region of interest.}

\item{mask}{SpatVector object used. Default = NULL.}

\item{raster_base}{SpatRaster representing the grid of the
raster layers used, which will be used for plotting purposes.}

\item{PCA_results}{results of principal component analysis performed with
values from raster layers used. Default = NULL.}
}
\value{
An S3 object of class \code{master_matrix}.
}
\description{
Constructor of S3 objects of class master_matrix
}
