% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{mx}
\alias{mx}
\title{Example of spatial polygon for a region of interest}
\format{
SpatVector
}
\description{
An object of class SpatVector.
}
\examples{
mx <- terra::vect(system.file("extdata/mx.gpkg", package = "biosurvey"))

mx
}
