% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{PAM_subset}
\alias{PAM_subset}
\alias{new_PAM_subset}
\title{Constructor of S3 objects of class PAM_subset}
\usage{
new_PAM_subset(PAM = new("SpatialPolygonsDataFrame"), PAM_indices = NULL,
               PAM_selected_sites_random = NULL, PAM_selected_sites_G = NULL,
               PAM_selected_sites_E = NULL, PAM_selected_sites_EG = NULL)
}
\arguments{
\item{PAM}{a SpatialPolygonsDataFrame object associated to information about
species presence and absence in a goegraphic grid.}

\item{PAM_indices}{list of indices derived from a PAM. Default = NULL.}

\item{PAM_selected_sites_random}{subset of \code{PAM} for sites derived from
random selection. Default = NULL.}

\item{PAM_selected_sites_G}{subset of \code{PAM} for sites derived from
selection considering geographic distances. Default = NULL.}

\item{PAM_selected_sites_E}{subset of \code{PAM} for sites derived from
selection considering environmental distances. Default = NULL.}

\item{PAM_selected_sites_EG}{subset of \code{PAM} for sites derived from
selection considering environment and geography. Default = NULL.}
}
\value{
An object of class \code{PAM_subset}.
}
\description{
Constructor of S3 objects of class PAM_subset
}
