% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_HD.R
\name{plot_matrix_HD}
\alias{plot_matrix_HD}
\title{Creates a plot to explore two of the environmental predictors in
geographic and environmental spaces.}
\usage{
plot_matrix_HD(master, variable_1, variable_2, col_variable1 = NULL,
               col_variable2 = NULL, col_points = NULL, col_density = NULL)
}
\arguments{
\item{master}{a master_matrix object derived from function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{uniformG_selection}}, \code{\link{uniformE_selection}}
or \code{EG_selection}.}

\item{variable_1}{(character or numeric) name or position of the first
variable (X axis) to be used to create blocks.}

\item{variable_2}{(character or numeric) name or position of the second
variable (Y axis) to be used to create blocks (must be different from the
first one).}

\item{col_variable1}{(character) a color palette, i.e. a vector of n contiguous
colors generated by functions like \code{\link{rainbow}},
\code{\link{heat.colors}}, \code{\link{topo.colors}},
\code{\link{bpy.colors}} or one of your own making, perhaps using
\code{\link{colorRampPalette}}. Default = NULL}

\item{col_variable2}{(character) a color palette, i.e. a vector of n contiguous
colors generated by functions like \code{\link{rainbow}},
\code{\link{heat.colors}}, \code{\link{topo.colors}},
\code{\link{bpy.colors}} or one of your own making, perhaps using
\code{\link{colorRampPalette}}. Default = NULL}

\item{col_points}{(character) a color palette, i.e. a vector of n contiguous
colors generated by functions like \code{\link{rainbow}},
\code{\link{heat.colors}}, \code{\link{topo.colors}},
\code{\link{bpy.colors}} or one of your own making, perhaps using
\code{\link{colorRampPalette}}. Default = NULL}

\item{col_density}{(character) a color palette, i.e. a vector of n contiguous
colors generated by functions like \code{\link{rainbow}},
\code{\link{heat.colors}}, \code{\link{topo.colors}},
\code{\link{bpy.colors}} or one of your own making, perhaps using
\code{\link{colorRampPalette}}. Default = NULL}
}
\value{
A multipanel plot showing two of the environmental preditors in the region of
interest in both spaces, geographic and environmental.
}
\description{
Creates a four panels plot with the information of two of the
environmental predictors in the region of interest. The two top panels contain
the information in geographic space (one predictor per panel) and in the two
bottom panels contain the information in environmental space (point cloud on
the left side and points density on the right side).
}
\examples{
# Data
data("m_matrix", package = "biosurvey")

colnames(m_matrix$master_matrix)

# Plot
plot_matrix_HD(m_matrix, variable_1 = "PC1", variable_2 = "PC2")
}
