% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SAC.R
\name{plot_SAC}
\alias{plot_SAC}
\title{Plotting lists of species accumulation curves}
\usage{
plot_SAC(
  SAC_selected_sites,
  col_mean = "blue",
  col_CI = "lightblue",
  alpha_mean = 0.7,
  alpha_CI = 0.2,
  xlab = "Number of sites",
  ylab = "Species",
  line_for_multiple = TRUE,
  ...
)
}
\arguments{
\item{SAC_selected_sites}{nested list of "\code{specaccum}" objects obtained
with function \code{\link{selected_sites_SAC}}.}

\item{col_mean}{(character) color for mean value of curve; default = "blue".}

\item{col_CI}{(character) color for confidence interval region for the curve;
default = "lightblue".}

\item{alpha_mean}{(numeric) alpha level for line representing the mean, values
from 0 to 1; default = 0.7. Values close to 0 increase transparency.}

\item{alpha_CI}{(numeric) alpha level for the region representing the confidence
interval; default = 0.2.}

\item{xlab}{(character) label for x axis of plot; default = "Number of sites".}

\item{ylab}{(character) label for y axis of plot; default = "Species".}

\item{...}{other arguments to be passed to plot method for objects of class
"\code{specaccum}".}
}
\value{
A plot of species "\code{specaccum}" objects. Multiple panels will be plotted
if \code{SA_selected_sites} list contains more than one element.
}
\description{
Creates species accumulation curve plots (one or multiple panels)
from information contained in lists.
}
\examples{
# data
data("b_pam", package = "biosurvey")
data("m_selection", package = "biosurvey")

# Subsetting base PAM according to selections
sub_pam_all <- subset_PAM(b_pam, m_selection, selection_type = "all")

SACs <- selected_sites_SAC(PAM_subset = sub_pam_all, selection_type = "all")

# plotting
plot_SAC(SACs)
}
