% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{preselected}
\alias{preselected}
\title{Example of a data.frame of preselected sites}
\format{
data.frame:
\describe{
\item{Site}{name of preselected sites}
\item{Longitude}{x coordinates}
\item{Latitude}{y coordinates}
}
}
\usage{
preselected
}
\description{
A data.frame with 5 rows and three columns: "Site", "Longitude",
and "Latitude".
}
\examples{
data("preselected", package = "biosurvey")

print(preselected)
}
\keyword{datasets}
