% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print}
\alias{print}
\alias{print.master_matrix}
\alias{print,master_matrix-method}
\alias{print,master_selection-method}
\alias{print,base_PAM-method}
\alias{print,PAM_subset-method}
\alias{print.master_selection}
\alias{print.base_PAM}
\alias{print.PAM_subset}
\alias{print.PAM_CS}
\title{Print a short version of elements in master objects}
\usage{
\method{print}{master_matrix}(x, ...)

\method{print}{master_selection}(x, ...)

\method{print}{base_PAM}(x, ...)

\method{print}{PAM_subset}(x, ...)

\method{print}{PAM_CS}(x, ...)
}
\arguments{
\item{x}{object of class master_matrix, master_selection, base_PAM, or
PAM_subset.}

\item{...}{further arguments to be passed to or from other methods. Ignored
in these functions.}
}
\description{
Print a short version of elements in master objects
}
