% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{purplow}
\alias{purplow}
\alias{darkros}
\alias{daright}
\alias{greeple}
\alias{bluered}
\title{Simple color palettes}
\usage{
# sequential palettes
purplow(n)

darkros(n)

daright(n)

# diverging palettes
greeple(n)

bluered(n)
}
\arguments{
\item{n}{number of colors to be in the palette.}
}
\value{
A character vector of hex color codes.
}
\description{
Simple color palettes
}
\examples{
purplow(3)
darkros(3)
daright(3)
greeple(3)
bluered(3)
}
