% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_matrix_cb.R
\name{randomize_matrix_cb}
\alias{randomize_matrix_cb}
\title{Randomize matrix using the curve ball algorithm}
\usage{
randomize_matrix_cb(matrix)
}
\arguments{
\item{matrix}{matrix to be randomized.}
}
\value{
A randomized matrix.
}
\description{
Randomize a matrix using the curve ball algorithm. This keeps the sum of
rows and columns constant.
}
\details{
The curve ball algorithm was described by Strona et al.
(2014)\link{https://www.nature.com/articles/ncomms5114}.
}
