% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam_helpers.R
\name{refill_PAM_indices}
\alias{refill_PAM_indices}
\title{Helper to refill a list of PAM indices with new or more results}
\usage{
refill_PAM_indices(initial_index_list, new_index_list)
}
\arguments{
\item{initial_index_list}{list of PAM indices to be refill. Indices present
in this list and absent in \code{new_index_list} are maintained.}

\item{new_index_list}{list of PAM indices to be used to refill
\code{initial_index_list}. New indices are included in the resulting list.
Indices present in both lists are updated using the values of this list.}
}
\value{
A list of PAM indices containing old and new values for its indices.
}
\description{
Helper to refill a list of PAM indices with new or more results
}
