% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam_helpers.R
\name{rlist_2data}
\alias{rlist_2data}
\title{Creates a data frame of species' references from a list of raster layers}
\usage{
rlist_2data(raster_list)
}
\arguments{
\item{raster_list}{list of RasterLayer objects. Each raster layer must be named
as the species that it represents, and values in each layer must be
1 (presence) and 0 (absence).}
}
\value{
A data frame of species geographic records derived from values of presence
in each layer from the ist of raster layers.
}
\description{
Creates a data frame of species' references that contains longitude,
latitude, and species name, using a list of raster layers as input. Useful when
raster layers have distinct extent or resolution.
}
