% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_master.R
\name{save_master_matrix}
\alias{save_master_matrix}
\alias{read_master_matrix}
\title{Functions to save and read biosurvey objects}
\usage{
save_master_matrix(master_matrix, file_name, verbose = TRUE)

read_master_matrix(file_name)
}
\arguments{
\item{master_matrix}{object derived from function prepare_master_matrix.}

\item{file_name}{path (with object name) for saving or reading the master matrix.}

\item{verbose}{whether or not to print messages about the process.
Default = TRUE}
}
\value{
if verbose = TRUE, returns a message indicating the path where the object was saved.

a master matrix object with unwrapped spatial objects.
}
\description{
Functions to save and read biosurvey objects
}
\examples{
mm <- read_master_matrix(file_name = "data/m_matrix.rds")
}
