% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pam.R
\name{save_PAM}
\alias{save_PAM}
\alias{read_PAM}
\title{Functions to save and read master objects}
\usage{
save_PAM(PAM, file_name, verbose = TRUE)

read_PAM(file_name)
}
\arguments{
\item{PAM}{object of class base_PAM or PAM_subset derived from the functions
\code{\link{prepare_base_PAM}} or \code{\link{subset_PAM}}.}

\item{file_name}{(character) name for the file to save or read the PAM
object (includes extension ".rds").}

\item{verbose}{whether or not to print messages about the process.
Default = TRUE.}
}
\value{
If \code{verbose} = TRUE, a message indicating the path where the object
was saved.

an object of class base_PAM with its spatial object unwrapped.
}
\description{
Functions to save and read master objects
}
\examples{
b_pam <- read_PAM(system.file("extdata/b_pam.rds",
                              package = "biosurvey"))
}
