% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam_helpers.R
\name{selected_sites_PAM}
\alias{selected_sites_PAM}
\title{Helper to subset PAM according to selected sites}
\usage{
selected_sites_PAM(selected_sites, base_PAM)
}
\arguments{
\item{selected_sites}{list of selected sites. See any of the main functions
to perform such a selection: \code{\link{random_selection}},
\code{\link{uniformG_selection}}, \code{\link{uniformE_selection}}, or
\code{\link{EG_selection}}.}

\item{base_PAM}{object of class base_PAM obtained using the function
\code{\link{base_PAM}}.}
}
\value{
A list of selected site data.frames with information of PAM added as
additional columns.
}
\description{
Helper to subset PAM according to selected sites
}
