% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{sp_layers}
\alias{sp_layers}
\title{Example of stack of layers of suitable and unsuitable conditions for species}
\format{A RasterStack with 109 rows, 182 columns, 19838 cells, and 5 layers:
\describe{
\item{RasterLayer}{suitable (1) and unsuitable (0) conditions}
}}
\description{
An object of class RasterStack containing information about suitable and
unsuitable conditions for five species.
}
\examples{
sp_layers <- raster::stack(system.file("extdata/sp_layers.tif",
                           package = "biosurvey"))

sp_layers
}
