% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sp_occurrences}
\alias{sp_occurrences}
\title{Occurrence records for the species Parides gundlachianus}
\format{
A data.frame with 19 rows and 3 columns.
\describe{
\item{name}{character, species scientific name.}
\item{longitude}{numeric, longitude values.}
\item{latitude}{numeric, latitude values.}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
sp_occurrences
}
\description{
Dataset containing geographic coordinates of a Cuban
butterfly.
}
\examples{
data("sp_occurrences", package = "biosurvey")

head(sp_occurrences)
}
\keyword{datasets}
