% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{species_data}
\alias{species_data}
\title{Example of species ranges as a SpatVector}
\format{
SpatVector
}
\description{
An object of class SpatVector.
}
\examples{
species_data <- terra::vect(system.file("extdata/species_data.gpkg",
                                        package = "biosurvey"))

species_data
}
