% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary}
\alias{summary}
\alias{summary.master_matrix}
\alias{summary,master_matrix-method}
\title{Summary of attributes and results}
\usage{
\method{summary}{master_matrix}(object)
}
\arguments{
\item{object}{object of class master_matrix.}
}
\value{
A printed summary.
}
\description{
Summary of attributes and results
}
