% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{variables}
\alias{variables}
\title{Example of variables to be used for preparing a master matrix}
\format{
A SpatRaster with 109 rows, 190 columns, 20710 cells, and 6 layers:
\describe{
\item{Temperature}{temperature, in degrees Celsius times 10.}
\item{Precipitation}{precipitation, in millimeters.}
}
}
\source{
\url{https://www.worldclim.org}
}
\description{
A dataset containing raster variables for an area that is relevant for used
in examples included in the package \code{\link{biosurvey}}.
}
\examples{
variables <- terra::rast(system.file("extdata/variables.tif",
                                     package = "biosurvey"))

terra::plot(variables[[1]])
}
