% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{b_pam}
\alias{b_pam}
\title{Example of object obtained from using the function base_PAM}
\format{
A list of 2 elements:
\describe{
\item{PAM}{SpatialPolygonsDataFrame wiht 306 features}
\item{PAM_statistics}{empty element (NULL)}
}
}
\usage{
b_pam
}
\description{
A list of 2 elements (PAM, PAM_statistics). See functions \code{\link{base_PAM}}.
}
\examples{
data("b_pam", package = "biosurvey")

summary(b_pam)
}
\keyword{datasets}
