% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{b_pam}
\alias{b_pam}
\title{Example of object obtained from using the function base_PAM}
\format{
A list of 2 elements:
\describe{
\item{PAM}{SpatVector with 306 features}
\item{PAM_indices}{a list of 11 elements}
}
}
\description{
An S3 object of class base_PAM. See functions \code{\link{prepare_base_PAM}}.
}
\examples{
b_pam <- read_PAM(file_name = "data/b_pam.rds")

print(b_pam)
}
