% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{base_PAM}
\alias{base_PAM}
\alias{new_base_PAM}
\title{Constructor of S3 objects of class base_PAM}
\usage{
new_base_PAM(PAM = new("SpatialPolygonsDataFrame"), PAM_indices = NULL)
}
\arguments{
\item{PAM}{SpatialPolygonsDataFrame object associated with information about
species presence and absence in a geographic grid.}

\item{PAM_indices}{list of indices derived from a PAM. Default = NULL.}
}
\value{
An object of class \code{base_PAM}.
}
\description{
Constructor of S3 objects of class base_PAM
}
