% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_pam.R
\name{base_pam}
\alias{base_pam}
\title{Prepare a presence-absence matrix (PAM)}
\usage{
base_pam(data, format = NULL, master_matrix, cell_size,
         complete_cover = TRUE)
}
\arguments{
\item{data}{species geographic ranges to be used to create a presence-absence
matrix (PAM). This argument can be: RasterStack, RasterBrick, data.frame,
list, SpatialPolygonsDataFrame, SpatialPointsDataFrame, or character. See
details for description of characteristics of each option.}

\item{format}{(character) if \code{data} is a character, available formats are:
"shp", "gpkg", "GTiff", and "ascii".}

\item{master_matrix}{object derived from function \code{\link{master_matrix}}.
Optionally, if master_matrix is not necessary, a list containing an object of
class SpatialPolygonsDataFrame, representinng the region of interest, can be
used. The name of this element in the list must be "polygon". For instance:
\code{my_list <- list(polygon = YOUR_SpatialPolygonsDataFrame)}.}

\item{cell_size}{(numeric) resolution for grid (single number or vector of two
numbers) in kilometers (km).}

\item{complete_cover}{(logical) whether or not to include cells of grid
partially overlapped with the geographic region of interest contined in
\code{master_matrix}. Default = TRUE.}
}
\value{
A presence-absence matrix (PAM) of class base_pam for the region of interest
associated with a SpatialPolygonsDataFrame, as in a grid of \code{cell_size}
resolution. Each grid cell is related to a specific ID and longitude and
latitude coordinates. Presence (1) and absence (0) values for each species
in every cell of the PAM are included as apart of the data frame of the
SpatialPolygonsDataFrame.
}
\description{
Prepare a presence-absence matrix (PAM) in which all initial points
of interest (rows) will have a representation of the species present in such
areas (columns). Initial points of interest will be represented by an ID, and
longitude and latitude coordinates.
}
\examples{
# Data
data("m_matrix", package = "biosurvey")
data("species_data", package = "biosurvey")

# Create base_pam
b_pam <- base_pam(data = species_data, master_matrix = m_matrix, cell_size = 100)
sp::plot(b_pam$PAM)
summary(b_pam$PAM@data[, 1:6])
}
