% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosurvey.R
\docType{package}
\name{biosurvey}
\alias{biosurvey}
\title{biosurvey: Tools for Biological Survey Planning}
\description{
biosurvey is a collection of tools that allow users to select sampling sites
increasing efficiency of biodiversity survey systems by considering the
relationship of environmental and geographic conditions in a region.
}
\section{Main functions in biosurvey}{

\code{\link{base_PAM}}, \code{\link{block_sample}},
\code{\link{compare_SAC}}, \code{\link{EG_selection}},
\code{\link{explore_data_EG}}, \code{\link{make_blocks}},
\code{\link{master_matrix}}, \code{\link{PAM_indices}},
\code{\link{plot_blocks_EG}}, \code{\link{plot_SAC}},
\code{\link{plot_sites_EG}}, \code{\link{random_selection}},
\code{\link{selected_sites_SAC}}, \code{\link{subset_PAM}},
\code{\link{uniformE_selection}}, \code{\link{uniformG_selection}}

Other functions (important helpers)

\code{\link{closest_to_centroid}}, \code{\link{files_2data}},
\code{\link{find_clusters}}, \code{\link{find_modes}},
\code{\link{grid_from_region}}, \code{\link{match_rformat}},
\code{\link{PAM_from_table}}, \code{\link{point_sample}},
\code{\link{point_sample_cluster}}, \code{\link{point_thinning}},
\code{\link{rlist_2data}}, \code{\link{selected_sites_PAM}},
\code{\link{spdf_2data}}, \code{\link{stack_2data}},
\code{\link{unimodal_test}}, \code{\link{wgs84_2aed_laea}}
}

