% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EG_selection.R
\name{EG_selection}
\alias{EG_selection}
\title{Selection of survey sites maximizing uniformity in environmental space
considering geographic structure}
\usage{
EG_selection(master, variable_1, variable_2, select_point = "E_centroid",
             cluster_method = "hierarchical", sample_for_distance = 250)
}
\arguments{
\item{master}{a master_matrix object derived from function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{random_selection}}, \code{\link{uniformG_selection}},
or \code{\link{uniformE_selection}}.}

\item{variable_1}{(character or numeric) name or position of the first
variable (X-axis).}

\item{variable_2}{(character or numeric) name or position of the second
variable (Y-axis).}

\item{select_point}{(character) How or which point will be selected. Three
options are available: "random", "E_centroid", "G_centroid". E_ or G_ centroid
indicate that the point(s) closests to the respective centroid will be selected.
Default = "E_centroid".}

\item{cluster_method}{(character) name of the method to be used for detecting
clusters. Options are "hierarchical" and "k-means"; default = "hierarchical".
See details in \code{\link{find_clusters}}}

\item{sample_for_distance}{(numeric) sample to be considered when measuring
the geographic distances among points in the blocks of environmental points.
Default = 250.}
}
\value{
A master_selection object (S3) with an aditional element called
selected_sites_EG containing one set of selected sites.
}
\description{
Selection of sites to be sampled in a survey, with the goal of
maximizing uniformity of points in environmental, but considering geographic
patterns of data. Similar environments that have a disjoint pattern are selected
twice (two survey sites are placed so they consider the biggest geographic
clusters).
}
