% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EG_selection.R
\name{EG_selection}
\alias{EG_selection}
\title{Selection of survey sites maximizing uniformity in environmental space
considering geographic structure}
\usage{
EG_selection(master, variable_1, variable_2, n_blocks, initial_distance,
             increase, replicates = 10, max_n_samples = 1,
             select_point = "E_centroid", cluster_method = "hierarchical",
             sample_for_distance = 250, set_seed = 1)
}
\arguments{
\item{master}{a master_matrix object derived from function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{random_selection}}, \code{\link{uniformG_selection}},
or \code{\link{uniformE_selection}}.}

\item{variable_1}{(character or numeric) name or position of the first
variable (X-axis).}

\item{variable_2}{(character or numeric) name or position of the second
variable (Y-axis).}

\item{n_blocks}{(numeric) number of blocks to be selected from all existent
blocks in \code{master$master_matrix}.}

\item{initial_distance}{(numeric) euclidean distance to be used for a first
process of thinning and detection of remaining points.}

\item{increase}{(numeric) value to be added to \code{initial_distance} until
reaching the number of \code{n_blocks}.}

\item{replicates}{(numeric) number of thinning replicates performed to select
blocks uniformly. Default = 10.}

\item{max_n_samples}{(numeric) maximum number of samples to be chosen after
performing all thinning \code{replicates}. Default = 1.}

\item{select_point}{(character) How or which point will be selected. Three
options are available: "random", "E_centroid", "G_centroid". E_ or G_ centroid
indicate that the point(s) closets to the respective centroid will be selected.
Default = "E_centroid".}

\item{cluster_method}{(character) name of the method to be used for detecting
clusters. Options are "hierarchical" and "k-means"; default = "hierarchical".
See details in \code{\link{find_clusters}}}

\item{sample_for_distance}{(numeric) sample to be considered when measuring
the geographic distances among points in the blocks of environmental points.
Default = 250.}

\item{set_seed}{(numeric) integer value to specify a initial seed. Default = 1.}
}
\value{
A master_selection object (S3) with an additional element called
selected_sites_EG containing one or more sets of selected sites depending on
\code{max_n_samples}.
}
\description{
Selection of sites to be sampled in a survey, with the goal of
maximizing uniformity of points in environmental, but considering geographic
patterns of data. Similar environments that have a disjoint pattern are selected
twice (two survey sites are placed so they consider the biggest geographic
clusters).
}
\examples{
# Data
data("m_matrix", package = "biosurvey")

# Making blocks for analysis
m_blocks <- make_blocks(m_matrix, variable_1 = "PC1", variable_2 = "PC2",
                        n_cols = 10, n_rows = 10, block_type = "equal_area")

# Checking column names
colnames(m_blocks$master_matrix)

# Selecting sites uniformly in E and G spaces
EG_sel <- EG_selection(master = m_blocks, variable_1 = "PC1", variable_2 = "PC2",
                       n_blocks = 10, initial_distance = 1.5, increase = 0.1,
                       replicates = 1, max_n_samples = 1,
                       select_point = "E_centroid",
                       cluster_method = "hierarchical",
                       sample_for_distance = 100)

head(EG_sel$selected_sites_EG[[1]])
dim(EG_sel$selected_sites_EG[[1]])
}
