% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionEG_helpers.R
\name{find_modes}
\alias{find_modes}
\title{Find modes in a multimodal distribution}
\usage{
find_modes(density)
}
\arguments{
\item{density}{an object of class density obtained using the function
\code{\link{density}}.}
}
\value{
A data.frame containing the values corresponding to the modes and the
density for those particular values.
}
\description{
Find modes in a multimodal distribution of values based on the
density of such values.
}
\examples{
# Data
data("dist_list", package = "biosurvey")

dens <- density(dist_list$`12`)

# Finding modes
modes <- find_modes(density = dens)
modes
}
