% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{m_matrix}
\alias{m_matrix}
\title{Example of a master_matrix object with no preselected sites}
\format{
A list of 6 elements:
\describe{
\item{data_matrix}{data.frame with 6276 rows and 10 columns}
\item{preselected_sites}{NULL}
\item{region}{object of class SpatVector}
\item{mask}{NULL}
\item{raster_base}{object of class SpatRaster}
\item{PCA_results}{list of length 4}
}
}
\description{
An S3 object of class master_matrix.
See function \code{\link{prepare_master_matrix}}.
}
\examples{
m_matrix <- read_master(system.file("extdata/m_matrix.rds",
                                    package = "biosurvey"))

print(m_matrix)
}
