% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{m_matrix_pre}
\alias{m_matrix_pre}
\title{Example of a master_matrix object containing preselected sites}
\format{
A list of 6 elements:
\describe{
\item{data_matrix}{data.frame with 6276 rows and 10 columns}
\item{preselected_sites}{data.frame with 5 rows and 11 columns}
\item{region}{object of class SpatVector}
\item{mask}{NULL}
\item{raster_base}{object of class SpatRaster}
\item{PCA_results}{list of length 4}
}
}
\description{
A S3 object of class master_matrix.
See function \code{\link{prepare_master_matrix}}.
}
\examples{
m_matrix_pre <- read_master(system.file("extdata/m_matrix_pre.rds",
                                        package = "biosurvey"))

print(m_matrix_pre)
}
