% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{m_selection}
\alias{m_selection}
\title{Example of a master_selection object from using functions for selecting sites}
\format{
A list of 10 elements:
\describe{
\item{data_matrix}{data.frame with 6276 rows and 10 columns}
\item{preselected_sites}{NULL}
\item{region}{object of class SpatialPolygons*}
\item{mask}{NULL}
\item{raster_base}{object of class SpatialPolygonsDataFrame}
\item{PCA_results}{list of length 5}
\item{selected_sites_random}{list with one data.frame}
\item{selected_sites_G}{list with one data.frame}
\item{selected_sites_E}{list with one data.frame}
\item{selected_sites_EG}{NULL}
}
}
\usage{
m_selection
}
\description{
An S3 object of class master_selection. See functions
\code{\link{uniformE_selection}}, \code{\link{uniformG_selection}},
\code{\link{random_selection}}, or \code{\link{EG_selection}}.
}
\examples{
data("m_selection", package = "biosurvey")

print(m_selection)
}
\keyword{datasets}
