% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_matrix.R
\name{master_matrix}
\alias{master_matrix}
\title{Prepare a base object to perform further analyses}
\usage{
master_matrix(region, variables, do_pca = FALSE, center = TRUE, scale = FALSE,
              variables_in_matrix = NULL)
}
\arguments{
\item{region}{SpatialPolygonsDataFrame of the region of interest.}

\item{variables}{RasterStack or RasterBrick of environmental variables.}

\item{do_pca}{(logical) whether or not to perfor a Principal Component Analysis.
Default = FALSE.}

\item{center}{(logical) whether or not to center variables. Argument to be passed
to the function \code{\link[stats]{prcomp}}. Default = TRUE.}

\item{scale}{(logical) whether or not to scale the variables. Recommended when
variables are in different units. Argument to be passed to the function
\code{\link[stats]{prcomp}}. Default = FALSE.}

\item{variables_in_matrix}{(character) name of variables to include in matrix.
If NULL (the default) all variables will be inluded.}
}
\value{
An S3 object of class master_matrix containing the following elements:
\itemize{
\item master_matrix: a date.frame with information about geographic location of
raster cells, initial environmental data, and if \code{do_pca} is TRUE,
the first two principal components derived from original data.
\item polygon: a SpatialPolygonsDataFrame representing the region of interest.
\item raster_base: a raster layer for the region of interest with a single value,
to be used for plotting purposes.
\item PCA_results: if \code{do_pca} is TRUE, other results from principal
component analysis. If FALSE, PCA_results element of the object is NULL.
}
}
\description{
Prepare an S3 object that will serve as the base to perform all
further analyses. This object will contain geographic and environmental
information that will be used to characterize the region of interest.
}
\examples{
# Data
data("mx", package = "biosurvey")
variables <- raster::stack(system.file("extdata/variables.tif",
                                       package = "biosurvey"))

# Create master matrix object
m_matrix <- master_matrix(region = mx, variables = variables, do_pca = TRUE,
                          center = TRUE, scale = TRUE)
}
