% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_matrix.R
\name{master_matrix}
\alias{master_matrix}
\title{Prepare a base object to perform further analyses}
\usage{
master_matrix(region, variables, do_pca = FALSE, center = TRUE, scale = FALSE,
              variables_in_matrix = NULL)
}
\arguments{
\item{region}{SpatialPolygonsDataFrame of the region of interest.}

\item{variables}{RasterStack or RasterBrick of environmental variables.}

\item{do_pca}{(logical) whether or not to perfor a Principal Component Analysis.
Default = FALSE.}

\item{center}{(logical) whether or not to center variables. Argument to be passed
to the function \code{\link[stats]{prcomp}}. Default = TRUE.}

\item{scale}{(logical) whether or not to scale the variables. Recommended when
variables are in different units. Argument to be passed to the function
\code{\link[stats]{prcomp}}. Default = FALSE.}

\item{variables_in_matrix}{(character) name of variables to include in matrix.
If NULL (the default) all variables will be inluded.}
}
\value{
A list containing the following elements:
\itemize{
\item A master matrix for the region of interest, with information
about geographic location of raster cells, initial environmental data,
and if \code{do_pca} is TRUE, the principal components derived from original data.
\item A spatial polygon representing the region of interest.
\item If \code{do_pca} is TRUE, other results from principal component analysis.
}
}
\description{
Prepare an S3 object that will serve as the base to perform all
further analyses. This object will contain geographic and environmental
information that will be used to characterize the region of interest.
}
