% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mx}
\alias{mx}
\title{Example of spatial polygon for a region of interest}
\format{
SpatialPolygonsDataFrame:
\describe{
\item{data}{data.frame wiht 1 row and 11 columns}
\item{polygons}{SpatialPolygons}
\item{proj4string}{object of class CRS}
}
}
\usage{
mx
}
\description{
An object of class SpatialPolygonsDataFrame.
}
\examples{
data("mx", package = "biosurvey")

mx
}
\keyword{datasets}
