% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM_plots.R
\name{plot_PAM_geo}
\alias{plot_PAM_geo}
\title{Plot of PAM indices in geography}
\usage{
plot_PAM_geo(PAM, index = "RI", col_pal = NULL, border = NULL, cex = 0.9)
}
\arguments{
\item{PAM}{an object of class base_PAM.}

\item{index}{(character) code for the index to be plotted. Options are: "RI",
"RIN", "DF", or "MCC". Default = "RI"}

\item{col_pal}{color palette function to be used in defining colors for the
\code{index} to be plotted. The default, NULL, uses \code{viridis::viridis}.}

\item{border}{color for cell borders of the PAM grid. The default, NULL, does
not plot any border.}

\item{cex}{(numeric) value by which plotting elements should be magnified
relative to the default. Default = 0.9}
}
\value{
A plot of \code{index} represented in geography.
}
\description{
Plot of PAM indices in geography
}
\examples{
data("b_pam", package = "biosurvey")

plot_PAM_geo(b_pam, index = "RI")
}
