% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_PAM_CS.R
\name{prepare_PAM_CS}
\alias{prepare_PAM_CS}
\title{Preparing data for Christen-Soberón diagrams}
\usage{
prepare_PAM_CS(PAM, exclude_column = NULL, id_column = NULL,
               significance_test = FALSE, randomization_iterations = 100,
               CL = 0.05, picante_iterations = NULL,
               keep_randomizations = FALSE, parallel = FALSE,
               n_cores = NULL)
}
\arguments{
\item{PAM}{matrix, data.frame, or base_PAM object containing information on
species presence and absence for a set of sites. Sites are organized in the
rows and species in the columns. See details.}

\item{exclude_column}{(optional) name or numeric index of columns to be
excluded. Default = NULL.}

\item{id_column}{(optional) name or numeric index of column containing the ID
of sites-cells of the PAM. Default = NULL.}

\item{significance_test}{(logical) whether to perform a test to detect
sites-cells that are statistically significant (i.e., the pattern detected
can be distinguished o random expectations). Default = FALSE.}

\item{randomization_iterations}{(numeric) number of iterations for the
randomization test used to calculate statistical significance. Default = 100.}

\item{CL}{(numeric) confidence limit to detect statistically significant
values. Default = 0.05.}

\item{picante_iterations}{(numeric) number of iterations to be used for each
matrix randomization process (to be done \code{randomization_iterations} times).
This process is done using the function \code{\link[picante]{randomizeMatrix}}
from the package \code{picante}. The default, NULL, uses \code{2 * sum(PAM)}.}

\item{keep_randomizations}{(logical) whether to keep a matrix with all values
from the randomization process. Default = FALSE.}

\item{parallel}{(logical) whether to perform analyses in parallel.
Default = FALSE.}

\item{n_cores}{(numeric) number of cores to be used when \code{parallel} =
TRUE. The default, NULL, uses available cores - 1.}
}
\value{
An S3 object of class PAM_CS if \code{PAM} is a matrix or data.frame,
otherwise, the base_PAM that includes the PAM_CS object as part of PAM_indices.
}
\description{
Preparation of data and details to create Christen- Soberón
diagrams.
}
\details{
Christen-Soberón diagrams are plots that allow explorations of patterns of
biodiversity in a region based on the data of presence-absence matrices. The
plots to be produced using the information prepared here are a modification of
those presented in ...
}
\examples{
# data
data("b_pam", package = "biosurvey")

# preparing data for CS diagram
pcs <- prepare_PAM_CS(PAM = b_pam)

summary(pcs$PAM_indices$CS_diagram)
}
