% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print}
\alias{print}
\alias{print.master_matrix}
\alias{print,master_matrix-method}
\alias{print,master_selection-method}
\alias{print.master_selection}
\title{Print a short version of elements in master objects}
\usage{
\method{print}{master_matrix}(object)

\method{print}{master_selection}(object)
}
\arguments{
\item{object}{object of class master_matrix or master_selection.}
}
\description{
Print a short version of elements in master objects
}
