% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_selection.R
\name{random_selection}
\alias{random_selection}
\title{Random selection of survey sites}
\usage{
random_selection(master, n_sites, n_samplings = 1, set_seed = 1)
}
\arguments{
\item{master}{a master_matrix object derived from function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{uniformG_selection}}, \code{\link{uniformE_selection}}
or \code{EG_selection}.}

\item{n_sites}{(numeric) number of sites to be selected from
\code{master_matrix} to be used as sites to be sampled in survey.}

\item{n_samplings}{(numeric) number of processes of selection, which will
turn into multiple options for a process of survey planning. Default = 1.}

\item{set_seed}{(numeric) integer value to specify a initial seed. Default = 1.}
}
\value{
A master_selection object (S3) with an aditional element called
selected_sites_random containing one or more sets of selected sites.
}
\description{
Random selection of sites to be sampled in a survey. Sites are
selected from a set of points that are provided in one of the arguments.
}
\examples{
# Data
data("m_matrix", package = "biosurvey")

r_selection <- random_selection(m_matrix, n_sites = 20, n_samplings = 5)
}
