% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_master.R
\name{save_master}
\alias{save_master}
\alias{read_master}
\title{Functions to save and read master objects}
\usage{
save_master(master, file_name, verbose = TRUE)

read_master(file_name)
}
\arguments{
\item{master}{object derived from the functions
\code{\link{prepare_master_matrix}}, \code{\link{random_selection}},
\code{\link{uniformG_selection}}, \code{\link{uniformE_selection}}, or
\code{uniformEG_selection}.}

\item{file_name}{(character) name for the file to save or read the master
object.}

\item{verbose}{whether or not to print messages about the process.
Default = TRUE.}
}
\value{
If \code{verbose} = TRUE, a message indicating the path where the object
was saved.

a master matrix object with unwrapped spatial objects.
}
\description{
Functions to save and read master objects
}
\examples{
mm <- read_master(file_name = "data/m_matrix.rds")
}
