% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sp_data}
\alias{sp_data}
\title{Example of a data frame of species' found in distint positions}
\format{data frame:
\describe{
\item{ID}{identifier of position}
\item{Species}{different species in the table}
}}
\usage{
sp_data
}
\description{
A data frame with two columns: "ID" and "Species" and 590 rows.
It is the output of the function \code{\link{spdf_2data}}.
}
\examples{
data("sp_data", package = "biosurvey")

summary(sp_data)
}
\keyword{datasets}
