% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{Example of species ranges as SpatialPolygonsDataFrame}
\format{
data.frame:
\describe{
\item{data}{data.frame wiht 25 rows and 1 column}
\item{polygons}{SpatialPolygons}
\item{proj4string}{object of class CRS}
}
}
\usage{
species_data
}
\description{
An object of class SpatialPolygonsDataFrame.
}
\examples{
data("species_data", package = "biosurvey")

species_data
}
\keyword{datasets}
