% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_pam.R
\name{subset_PAM}
\alias{subset_PAM}
\title{Data

data("m_matrix", package = "biosurvey")
data("species_data", package = "biosurvey")
Subset PAM according to selected sites}
\usage{
subset_PAM(base_PAM, master_selection, selection_type = "all")
}
\description{
Data

data("m_matrix", package = "biosurvey")
data("species_data", package = "biosurvey")
Subset PAM according to selected sites
}
\examples{
# Create base_pam
b_pam <- base_pam(data = species_data, master_matrix = m_matrix, cell_size = 100)

# Making blocks for analysis with master matrix
m_blocks <- make_blocks(m_matrix, variable_1 = "PC1", variable_2 = "PC2",
                        n_cols = 20, n_rows = 20, block_type = "equal_area")

# Selections
## Randomly
m_selection <- random_selection(m_blocks, n_sites = 20, n_samplings = 2)

## Uniformly in G space
m_selection <- uniformG_selection(m_selection, expected_points = 20,
                                  max_n_samplings = 1, initial_distance = 165,
                                  increase = 1, replicates = 5)

## Uniformly in E space
m_selection <- uniformE_selection(m_selection, variable_1 = "PC1", variable_2 = "PC2",
                                  selection_from = "block_centroids",
                                  expected_points = 15, max_n_samplings = 1,
                                  initial_distance = 1, increase = 0.1,
                                  replicates = 5)
}
