% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary}
\alias{summary}
\alias{summary.master_matrix}
\alias{summary,master_matrix-method}
\alias{summary,master_selection-method}
\alias{summary,base_PAM-method}
\alias{summary,PAM_subset-method}
\alias{summary.master_selection}
\alias{summary.base_PAM}
\alias{summary.PAM_subset}
\alias{summary.PAM_CS}
\title{Summary of attributes and results}
\usage{
\method{summary}{master_matrix}(object, ...)

\method{summary}{master_selection}(object, nrow = 6, ncol = 2, ...)

\method{summary}{base_PAM}(object, ...)

\method{summary}{PAM_subset}(object, ...)

\method{summary}{PAM_CS}(object, ...)
}
\arguments{
\item{object}{object of class master_matrix or master_selection.}

\item{...}{additional arguments affecting the summary produced. Ignored in
these functions.}

\item{nrow}{number of rows to be printed for selected_sites in a
master_selection object.}

\item{ncol}{number of columns to be printed for selected_sites in a
master_selection object.}
}
\value{
A printed summary.
}
\description{
Summary of attributes and results
}
