% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformE_selection.R
\name{uniformE_selection}
\alias{uniformE_selection}
\title{Selection of survey sites maximizing uniformity in environmental space}
\usage{
uniformE_selection(master, variable_1, variable_2, selection_from = "all_points",
                   expected_points, max_n_samples = 1,
                   initial_distance, increase, replicates = 10, set_seed = 1)
}
\arguments{
\item{master}{a master_matrix object derived from function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{random_selection}}, \code{\link{uniformG_selection}},
or \code{\link{EG_selection}}.}

\item{variable_1}{(character or numeric) name or position of the first
variable (X-axis).}

\item{variable_2}{(character or numeric) name or position of the second
variable (Y-axis).}

\item{selection_from}{(character) set of points to perfomr the selection from.
Two options are available, "all_points" or "block_centroids". The first option
picks the points from all points in the environmental cloud, and the second
one selects points only from centroids of environmental blocks. See
\code{\link{make_blocks}}. Default = "all_points".}

\item{expected_points}{(numeric) number of survey points (sites) to be selected.}

\item{max_n_samples}{(numeric) maximun number of samples to be chosen after
performing all thinning \code{replicates}. Default = 1.}

\item{initial_distance}{(numeric) euclidean distance to be used for a first
process of thinning and detection of remaining points.}

\item{increase}{(numeric) value to be added to \code{initial_distance} untill
reaching the number of \code{expected_points}.}

\item{replicates}{(numeric) number of thinning replicates. Default = 10.}

\item{set_seed}{(numeric) integer value to specify a initial seed. Default = 1.}
}
\value{
A master_selection object (S3) with an aditional element called
selected_sites_E containing one or more sets of selected sites.
}
\description{
Selection of sites to be sampled in a survey, with the goal of
maximizing uniformity of points in environmental space.
}
\examples{
# Data
data("m_matrix", package = "biosurvey")

# Making blocks for analysis
m_blocks <- make_blocks(m_matrix, variable_1 = "PC1",
                        variable_2 = "PC2", n_cols = 10, n_rows = 10,
                        block_type = "equal_area")

# Checking column names
colnames(m_blocks$master_matrix)

# Selecting sites uniformly in E space
selectionE <- uniformE_selection(m_blocks, variable_1 = "PC1", variable_2 = "PC2",
                                 selection_from = "block_centroids",
                                 expected_points = 15, max_n_samples = 1,
                                 initial_distance = 1, increase = 0.1,
                                 replicates = 5, set_seed = 1)
}
