% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformG_selection.R
\name{uniformG_selection}
\alias{uniformG_selection}
\title{Selection of survey sites maximizing uniformity in geography}
\usage{
uniformG_selection(master, expected_points, max_n_samples = 1,
                   initial_distance, increase, replicates = 10, set_seed = 1)
}
\arguments{
\item{master}{a master_matrix object derived from function
\code{\link{master_matrix}} or a master_selection object derived from functions
\code{\link{random_selection}}, \code{\link{uniformE_selection}},
or \code{\link{EG_selection}}.}

\item{expected_points}{(numeric) number of survey points (sites) to be selected.}

\item{max_n_samples}{(numeric) maximun number of samples to be chosen after
performing all thinning \code{replicates}. Default = 1.}

\item{initial_distance}{(numeric) distance in km to be used for a first
process of thinning and detection of remaining points.}

\item{increase}{(numeric) value to be added to \code{initial_distance} untill
reaching the number of \code{expected_points}.}

\item{replicates}{(numeric) number of thinning replicates. Default = 10.}

\item{set_seed}{(numeric) integer value to specify a initial seed. Default = 1.}
}
\value{
A master_selection object (S3) with an aditional element called
selected_sites_G containing one or more sets of selected sites.
}
\description{
Selection of sites to be sampled in a survey, with the goal of
maximizing uniformity of points in geographic space.
}
\examples{
# Data
data("m_matrix", package = "biosurvey")

# Selecting sites uniformly in G space
selectionG <- uniformG_selection(m_matrix, expected_points = 40, max_n_samples = 1,
                                 initial_distance = 145, increase = 1,
                                 replicates = 5, set_seed = 1)
}
