% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformG_selection.R
\name{uniformG_selection}
\alias{uniformG_selection}
\title{Selection of survey sites maximizing uniformity in geography}
\usage{
uniformG_selection(master_matrix, expected_points, max_n_samples = 1,
                   initial_distance, increase, replicates = 10, set_seed = 1)
}
\arguments{
\item{master_matrix}{object derived from function \code{\link{master_matrix}}.
Optionally, if master_matrix is not necessary, a list containing an object of
class data.frame with at least two columns represening two variables. The name
of this element in the list must be "master_matrix". For instance:
\code{my_list <- list(master_matrix = YOUR_data.frame)}.}

\item{expected_points}{(numeric) number of survey points (sites) to be selected.}

\item{max_n_samples}{(numeric) maximun number of samples to be chosen after
performing all thinning \code{replicates}. Default = 1.}

\item{initial_distance}{(numeric) distance in km to be used as for a first
process of thinning and detection of remaining points.}

\item{increase}{(numeric) value to be added to \code{initial_distance} untill
reaching the number of \code{expected_points}.}

\item{replicates}{(numeric) number of thinning replicates. Default = 10.}

\item{set_seed}{(numeric) integer value to specify a initial seed. Default = 1.}
}
\value{
The master_matrix list with an aditional element containing one or more sets
of selected sites.
}
\description{
Selection of sites to be sampled in a survey, with the goal of
maximizing uniformity of points in geographic space.
}
