% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{variables}
\alias{variables}
\title{Example of variables to be used for preparing a master matrix}
\format{
A RasterStack with 109 rows, 190 columns, 20710 cells, and 6 layers:
\describe{
\item{Temperature}{temperature, in degrees Celsius times 10.}
\item{Precipitation}{precipitation, in milimeters.}
}
}
\source{
\url{https://www.worldclim.org/data/index.html}
}
\description{
A dataset containing raster variables for an area that is relevant for used
in examples included in the package \code{\link{biosurvey}}.
}
\examples{
variables <- raster::stack(system.file("extdata/variables.tif",
                                       package = "biosurvey"))

raster::plot(variables[[1]])
}
